PIの計算を実装

+++ BigFloat.hpp
@@ -16,7 +16,7 @@ struct BigFloat
 
 int CompareAbs(BigFloat &A, BigFloat &B);
 void InitializeBigFloat(BigFloat &A, bool sign = POSI, int exponent = 0,
-                        int fraction = 0);
+                        long long fraction = 0);
 void changeExp(BigFloat &A, int exp);
 void AddBigFloat(BigFloat A, BigFloat B, BigFloat &C);
 void MulBigFloat(BigFloat A, BigFloat B, BigFloat &C);
